NO AUTO Strips atomic, hashing, temporal, date, index, neighbours, merge from full
