+++ core/src/main/java/apoc/atomic/Atomic.java
