+++ core/src/main/java/apoc/atomic/util/AtomicUtils.java
