+++ core/src/main/java/apoc/hashing/Fingerprinting.java
