+++ core/src/main/java/apoc/hashing/FingerprintingConfig.java
