+++ core/src/main/java/apoc/index/QueuePoisoningCollector.java
