+++ core/src/main/java/apoc/index/SchemaIndex.java
