+++ core/src/main/java/apoc/spatial/Distance.java
