+++ core/src/test/java/apoc/index/SchemaIndexTest.java
