+++ core/src/test/java/apoc/spatial/DistanceTest.java
