+++ core/src/test/java/apoc/spatial/SpatialTest.java
