+++ full/src/test/java/apoc/ApocSplitTest.java
@@ -25,14 +25,6 @@ public class ApocSplitTest {
         "apoc.trigger.resume",
         "apoc.util.sleep",
         "apoc.util.validate",
-        "apoc.merge.node.eager",
-        "apoc.merge.node",
-        "apoc.merge.nodeWithStats.eager",
-        "apoc.merge.nodeWithStats",
-        "apoc.merge.relationship",
-        "apoc.merge.relationshipWithStats",
-        "apoc.merge.relationship.eager",
-        "apoc.merge.relationshipWithStats.eager",
         "apoc.nodes.cycles",
         "apoc.nodes.link",
         "apoc.nodes.get",
@@ -85,20 +77,11 @@ public class ApocSplitTest {
         "apoc.do.when",
         "apoc.case",
         "apoc.do.case",
-        "apoc.atomic.add",
-        "apoc.atomic.subtract",
-        "apoc.atomic.concat",
-        "apoc.atomic.insert",
-        "apoc.atomic.remove",
-        "apoc.atomic.update",
         "apoc.search.nodeAllReduced",
         "apoc.search.nodeReduced",
         "apoc.search.multiSearchReduced",
         "apoc.search.nodeAll",
         "apoc.search.node",
-        "apoc.schema.assert",
-        "apoc.schema.nodes",
-        "apoc.schema.relationships",
         "apoc.load.jsonArray",
         "apoc.load.json",
         "apoc.load.jsonParams",
@@ -106,13 +89,7 @@ public class ApocSplitTest {
         "apoc.import.xml",
         "apoc.load.arrow.stream",
         "apoc.load.arrow",
-        "apoc.schema.properties.distinct",
-        "apoc.schema.properties.distinctCount",
         "apoc.log.stream",
-        "apoc.spatial.sortByDistance",
-        "apoc.spatial.geocodeOnce",
-        "apoc.spatial.geocode",
-        "apoc.spatial.reverseGeocode",
         "apoc.create.node",
         "apoc.create.addLabels",
         "apoc.create.setProperty",
@@ -134,18 +111,9 @@ public class ApocSplitTest {
         "apoc.create.uuids",
         "apoc.warmup.run",
         "apoc.stats.degrees",
-        "apoc.help",
-        "apoc.neighbors.tohop",
-        "apoc.neighbors.tohop.count",
-        "apoc.neighbors.byhop",
-        "apoc.neighbors.byhop.count",
-        "apoc.neighbors.athop",
-        "apoc.neighbors.athop.count");
+        "apoc.help");
 
     public static final List<String> FUNCTIONS_FROM_CORE = List.of(
-        "apoc.temporal.format",
-        "apoc.temporal.formatDuration",
-        "apoc.temporal.toZonedTemporal",
         "apoc.util.sha1",
         "apoc.util.sha256",
         "apoc.util.sha384",
@@ -177,18 +145,6 @@ public class ApocSplitTest {
         "apoc.path.slice",
         "apoc.path.combine",
         "apoc.path.elements",
-        "apoc.date.toYears",
-        "apoc.date.fields",
-        "apoc.date.field",
-        "apoc.date.currentTimestamp",
-        "apoc.date.format",
-        "apoc.date.toISO8601",
-        "apoc.date.fromISO8601",
-        "apoc.date.parse",
-        "apoc.date.systemTimezone",
-        "apoc.date.convert",
-        "apoc.date.convertFormat",
-        "apoc.date.add",
         "apoc.label.exists",
         "apoc.meta.cypher.isType",
         "apoc.meta.cypher.type",
@@ -197,13 +153,6 @@ public class ApocSplitTest {
         "apoc.diff.nodes",
         "apoc.cypher.runFirstColumnMany",
         "apoc.cypher.runFirstColumnSingle",
-        "apoc.hashing.fingerprint",
-        "apoc.hashing.fingerprinting",
-        "apoc.hashing.fingerprintGraph",
-        "apoc.schema.node.indexExists",
-        "apoc.schema.relationship.indexExists",
-        "apoc.schema.node.constraintExists",
-        "apoc.schema.relationship.constraintExists",
         "apoc.xml.parse",
         "apoc.version",
         "apoc.scoring.existence",
